Fix to form to mongo
Fixed code so that it is better using auth.ts. Also updated temp login to make it better and easier for code testing. Added loading status when login pushes to dashboard. Items added through creating a listing now correctly add to the correct user in mongodb, and correctly display in dashboard. Noticed an issue where files uploaded from the local machine don't appear in the dashboard correctly, will attempt to fix later on in a different branch.